requirements.txt
amqp==2.6.1
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==4.4.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Deprecated==1.2.13
Django==4.0.7
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-filter==22.1
django-timezone-field==5.0
djangorestframework==3.13.1
gevent==21.12.0
greenlet==1.1.2
importlib-metadata==4.12.0
kombu==4.6.11
Markdown==3.4.1
packaging==21.3
prompt-toolkit==3.0.30
pycparser==2.21
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.2.1
redis==4.3.4
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
vine==1.3.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
